test(fuzz): link harnesses to a slim object lib so ASan stops aborting pre-main (#193)
Every libFuzzer harness linked $<TARGET_OBJECTS:ra_core_hal>, which pulls in
tests/mocks/ra_sim_mmap.c. Its load-time MAP_FIXED constructor maps the
0xE0000000 SCB/MPU window, which lands in AddressSanitizer's shadow gap on
Linux, so the process aborted in the constructor before main() -- taking down
`make fuzz` for every harness that links the peripheral mock.

Restructure the fuzz build so each harness links only what it needs:

- Add ra_core_hal_fuzz: ra_core_hal MINUS ra_sim_mmap.c (tests/CMakeLists.txt),
  built only when RA_FUZZ=ON. The pure-computation harnesses link it and never
  run the MAP_FIXED constructor.
- The three harnesses that dereference real peripheral registers (canfd,
  usb_pal, ble_att) add ra_sim_mmap.c back on their own link line, compiled
  under ASan. ra_sim_mmap now skips the single shadow-gap region under ASan
  (ra_sim_region_mappable); every non-ASan host-test build maps it as before,
  so `make test` is unchanged.
- fuzz_log_sink.c installs a discard ra_log byte sink from a constructor, so no
  harness reaches the ITM registers at 0xE0000000 (ra_log reads them under
  RA_SIMULATOR_MODE absent a sink; canfd logs during init).
- Fix a pre-existing signed-overflow UB in fuzz_ra_stbtt.c (x1 - x0 on INT_MIN
  box coords), which UBSan aborts on once the harness actually runs.

ble_att is reclassified as an MMIO harness: ra_ble_host_init -> ra_ble_open
writes the BLE controller block at 0x40700000.

(cherry picked from commit 9ad784b379ab09ff7c3ff1a4d155a1f82c3b868a)

+++ tests/fuzz/fuzz_ra_stbtt.c
@@ -67,15 +67,26 @@ int LLVMFuzzerTestOneInput(const uint8_t* data, size_t size)
     int x1 = 0;
     int y1 = 0;
     stbtt_GetCodepointBitmapBox(&font, s_codepoints[i], scale, scale, &x0, &y0, &x1, &y1);
-    const int w = x1 - x0;
-    const int h = y1 - y0;
+    /* stb_truetype can hand back extreme box coordinates (down to INT_MIN)
+     * for a crafted font, so compute the extents in int64_t: the naive
+     * `x1 - x0` overflows int and is undefined behaviour. The bound below
+     * (<= k_fuzz_glyph_max_dim) then discards any implausible box. */
+    const int64_t w = (int64_t)x1 - (int64_t)x0;
+    const int64_t h = (int64_t)y1 - (int64_t)y0;
     if (w <= 0 || h <= 0 || w > k_fuzz_glyph_max_dim || h > k_fuzz_glyph_max_dim) {
       continue;
     }
     if ((size_t)w * (size_t)h > sizeof s_bitmap) {
       continue;
     }
-    stbtt_MakeCodepointBitmap(&font, s_bitmap, w, h, w, scale, scale, s_codepoints[i]);
+    stbtt_MakeCodepointBitmap(&font,
+                              s_bitmap,
+                              (int)w,
+                              (int)h,
+                              (int)w,
+                              scale,
+                              scale,
+                              s_codepoints[i]);
   }
   return 0;
 }
